Draw 3D cube_pyramid using opengl via C language
Using OpenGL in c, this program draws the red/green/blue 3D Cube and Pyramid. The cube is made of 6 quads , each having different colors and the hollow pyramid is made up of 4 triangle, with different colors on  each of the vertices. when compiling this program you must link to the OpenGL library and to the glut library.